lb_router
Module execute

Functions

_add_liquidity 🔒

Helper function to add liquidity.

_get_amounts_in 🔒

Helper function to return the amounts in.

_get_lb_pair_information 🔒

Helper function to return the address of the LBPair.

_get_pair 🔒

Helper function to return the address of the pair (v1 or v2, according to binStep).

_get_pairs 🔒

Helper function to return a list of pairs.

_remove_liquidity 🔒

Helper function to remove liquidity.

_safe_transfer 🔒

Helper function to transfer tokens to to.

_safe_transfer_from 🔒

Helper function to transfer tokens from from to to.

_safe_transfer_native 🔒

Helper function to transfer NATIVE to to.

_swap_exact_tokens_for_tokens 🔒

Helper function to swap exact tokens for tokens.

_swap_tokens_for_exact_tokens 🔒

Helper function to swap tokens for exact tokens

_w_native_deposit_and_transfer 🔒

Helper function to deposit and transfer WNative to to.

_w_native_withdraw_and_transfer 🔒

Helper function to withdraw and transfer WNative to to.

add_liquidity

Add liquidity while performing safety checks.

add_liquidity_native

Add liquidity with NATIVE while performing safety checks.

create_lb_pair

Create a liquidity bin LBPair for token_x and token_y using the factory.

remove_liquidity

Remove liquidity while performing safety checks.

remove_liquidity_native

Remove NATIVE liquidity while performing safety checks.

swap_exact_native_for_tokens

Swaps exact NATIVE for tokens while performing safety checks.

swap_exact_tokens_for_native

Swaps exact tokens for NATIVE while performing safety checks.

swap_exact_tokens_for_tokens

Swaps exact tokens for tokens while performing safety checks.

swap_native_for_exact_tokens

Swaps NATIVE for exact tokens while performing safety checks.

swap_tokens_for_exact_native

Swaps tokens for exact NATIVE while performing safety checks.

swap_tokens_for_exact_tokens

Swaps tokens for exact tokens while performing safety checks.

sweep

Unstuck tokens that are sent to this contract by mistake.

sweep_lb_token

Unstuck LBTokens that are sent to this contract by mistake.